Transformations started and polymorphism done better in the core ABC class
